move isis signal into test_init
