+++ tests/devices/test_init.py
@@ -1,7 +1,8 @@
 from unittest.mock import patch
 
 import pytest
-from ibex_bluesky_core.devices import compress_and_hex, dehex_and_decompress, get_pv_prefix
+
+from ibex_bluesky_core.devices import compress_and_hex, dehex_and_decompress, get_pv_prefix, isis_epics_signal_rw
 
 
 def test_can_dehex_and_decompress():
@@ -29,3 +30,12 @@ def test_cannot_get_pv_prefix():
         mock_getenv.return_value = None
         with pytest.raises(EnvironmentError):
             get_pv_prefix()
+
+
+def test_isis_epics_rw_signal_appends_correct_sp_suffix():
+    with patch("ibex_bluesky_core.devices.epics_signal_rw") as mock_epics_signal_rw:
+        read_pv = "TEST"
+        expected_sp_pv = f"{read_pv}:SP"
+        datatype = int
+        isis_epics_signal_rw(datatype=datatype, read_pv=read_pv)
+        mock_epics_signal_rw.assert_called_with(int, read_pv, expected_sp_pv, "")
